% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extr-package.R
\docType{package}
\name{extr-package}
\alias{extr}
\alias{extr-package}
\title{extr: Extinction Risk Estimation}
\description{
Estimates extinction risk from population time series under a drifted Wiener process using MLE and observation-error-and-autocovariance-robust estimators, with confidence intervals based on the w-z method.
}
\author{
\strong{Maintainer}: Hiroshi Hakoyama \email{hiroshi.hakoyama@gmail.com} (\href{https://orcid.org/0000-0001-7464-0754}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Hiroshi Hakoyama \email{hiroshi.hakoyama@gmail.com} (\href{https://orcid.org/0000-0001-7464-0754}{ORCID}) [copyright holder]
}

}
\keyword{internal}
