% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extr-package.R
\docType{package}
\name{extr-package}
\alias{extr}
\alias{extr-package}
\title{extr: Extinction Risk Estimation}
\description{
Estimates extinction probabilities from time series of population sizes based on a Wiener-drift process model. Implements the w-z method (Hakoyama, 2025) to improve the accuracy of confidence intervals over conventional approaches.
}
\author{
\strong{Maintainer}: Hiroshi Hakoyama \email{hiroshi.hakoyama@gmail.com} (\href{https://orcid.org/0000-0001-7464-0754}{ORCID}) [copyright holder]

}
\keyword{internal}
