% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ext_di.R
\name{print.ext_di}
\alias{print.ext_di}
\title{Print Extinction Risk Estimations}
\usage{
\method{print}{ext_di}(x, digits = 5, ...)
}
\arguments{
\item{x}{an object of class \code{"ext_di"} produced by
\code{\link{ext_di}}.}

\item{digits}{integer: Number of significant digits to display. Default is 5.}
}
\value{
Invisibly returns \code{obj} after printing formatted extinction risk
estimation results, including estimates, confidence intervals, data
summary, and input parameters.
}
\description{
Print Extinction Risk Estimations
}
\author{
Hiroshi Hakoyama, \email{hiroshi.hakoyama@gmail.com}
}
\keyword{internal}
